Extract cross-episode scan modal and window-optimize panel out of CueTemplatesPanel
CueTemplatesPanel.tsx carried CueWindowOptimizePanel (~170 lines) and
CueCrossEpisodeScanModal (~305 lines) inline. Move both to their own
files under components/ (matching where the app keeps extracted modals),
pull the shared modal/button style recipes and useEscape into
cueScanStyles, and factor the duplicated claim/poll query config into a
useScanQuery hook (queryKey, queryFn, rescanFn -> {data, scanning,
scanError, rescan}). The hook's scanning derivation drops the dead
phase==='results' guard since the query's enabled already gates it.
