+++ frontend/src/pages/feeds/CueTemplatesPanel.tsx
@@ -4,26 +4,26 @@ import { X, Play, Square } from 'lucide-react';
 import CollapsibleSection from '../../components/CollapsibleSection';
 import LoadingSpinner from '../../components/LoadingSpinner';
 import CueMarkModal from '../../components/CueMarkModal';
+import CueWindowOptimizePanel from '../../components/CueWindowOptimizePanel';
+import CueCrossEpisodeScanModal from '../../components/CueCrossEpisodeScanModal';
+import {
+  ghostBtn, primaryBtn, fieldCls, modalBackdrop, modalPanel, useEscape,
+} from '../../components/cueScanStyles';
 import {
   CUE_TYPE_OPTIONS,
-  crossEpisodeScan,
   cueTemplateAudioUrl,
   cueTemplateExportUrl,
   deleteCueTemplate,
   importCueTemplate,
   listCueTemplates,
-  optimizeCueWindow,
   previewCueTemplate,
   scanEpisodeCues,
   suggestCueThreshold,
   updateCueTemplate,
-  type CrossEpisodeCandidate,
-  type CrossEpisodeScanResponse,
   type CueScanResponse,
   type CueTemplate,
   type CueTemplateScope,
   type CueTemplateType,
-  type CueWindowOptimizeResponse,
   type ThresholdSuggestResponse,
 } from '../../api/cueTemplates';
 import { getCueFeedAdvisory } from '../../api/cueDetections';
@@ -32,29 +32,11 @@ import { getSettings } from '../../api/settings';
 import type { Feed } from '../../api/types';
 import type { Episode } from '../../api/types';
 import { formatTime } from '../../utils/adReviewHelpers';
-import { formatTimestamp } from '../../utils/format';
 
 const PICKER_PAGE_SIZE = 50;
 
-// Design-system recipes shared by the panel and its modals (match the app's
-// confirm/edit modals and form controls; theme-aware in dark mode).
-const ghostBtn = 'border border-border hover:bg-accent transition-colors';
-const primaryBtn = 'bg-primary text-primary-foreground hover:bg-primary/90 transition-colors';
-const fieldCls = 'rounded-lg border border-input bg-background text-foreground focus:outline-hidden focus:ring-2 focus:ring-ring';
-const modalBackdrop = 'fixed inset-0 z-50 flex items-center justify-center bg-black/50 p-4';
-const modalPanel = 'bg-card text-foreground rounded-lg border border-border shadow-xl';
-
 interface Props {
   slug: string;
-}
-
-// Close-on-Escape for the lightweight modals below.
-function useEscape(onClose: () => void) {
-  useEffect(() => {
-    const onKey = (e: KeyboardEvent) => { if (e.key === 'Escape') onClose(); };
-    window.addEventListener('keydown', onKey);
-    return () => window.removeEventListener('keydown', onKey);
-  }, [onClose]);
 }
 
 // Per-feed cue template management. Templates take precedence over the global
@@ -1048,500 +1030,6 @@ function CueScanModal({ slug, onClose }: CueScanModalProps) {
         )}
       </div>
     </div>
-  );
-}
-
-interface CueWindowOptimizePanelProps {
-  slug: string;
-  template: CueTemplate;
-  onClose: () => void;
-}
-
-// Inline before/after panel for the window optimizer (D2b). Mounting claims or
-// polls the background sweep (D1b claim/poll convention); Apply moves the
-// window via the template PATCH, which re-extracts blobs server-side.
-function CueWindowOptimizePanel({ slug, template, onClose }: CueWindowOptimizePanelProps) {
-  const queryClient = useQueryClient();
-  const [applyError, setApplyError] = useState<string | null>(null);
-
-  const queryKey = ['cue-window-optimize', slug, template.id];
-  const query = useQuery<CueWindowOptimizeResponse>({
-    queryKey,
-    queryFn: () => optimizeCueWindow(slug, template.id),
-    staleTime: Infinity,
-    refetchInterval: (q) =>
-      q.state.data?.status === 'scanning' ? 3000 : false,
-  });
-
-  // Collapse keeps nothing: drop the cached proposal once the panel unmounts
-  // (Discard, Apply, or toggling the row action). Removing after unmount also
-  // avoids an observer refetch re-claiming a scan server-side.
-  useEffect(() => {
-    return () => {
-      queryClient.removeQueries({ queryKey: ['cue-window-optimize', slug, template.id] });
-    };
-  }, [queryClient, slug, template.id]);
-
-  const data = query.data;
-  const { proposedStartS, proposedEndS, meanPeakScore, baselineMeanPeakScore, perEpisode, baselineWindow } = data ?? {};
-  const scanning = query.isLoading || data?.status === 'scanning';
-  // A thrown trigger error (e.g. 409 source original aged out) carries the
-  // server message; a saved worker error comes back in the payload.
-  const scanError = data?.status === 'error'
-    ? (data.error || 'Optimize failed.')
-    : query.error
-      ? (query.error instanceof Error ? query.error.message : 'Optimize failed.')
-      : null;
-  const ready = data?.status === 'ready'
-    && proposedStartS != null && proposedEndS != null && meanPeakScore != null;
-  const alreadyOptimal = ready
-    && proposedStartS === baselineWindow?.startS
-    && proposedEndS === baselineWindow?.endS;
-  const scoreDelta = ready && baselineMeanPeakScore != null
-    ? meanPeakScore - baselineMeanPeakScore
-    : null;
-
-  const applyMutation = useMutation({
-    mutationFn: (vars: { startS: number; endS: number }) =>
-      updateCueTemplate(template.id, {
-        sourceOffsetS: vars.startS,
-        // Round away float dust from end - start; the sweep works in 0.1s steps.
-        durationS: Math.round((vars.endS - vars.startS) * 1000) / 1000,
-      }),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['cue-templates', slug] });
-      onClose();
-    },
-    onError: (e) => setApplyError(e instanceof Error ? e.message : 'Apply failed'),
-  });
-
-  // Force a fresh server-side run; same fetchQuery idiom as the cross-episode
-  // modal's rescan.
-  const rescan = () => {
-    setApplyError(null);
-    queryClient.fetchQuery({
-      queryKey,
-      queryFn: () => optimizeCueWindow(slug, template.id, true),
-      staleTime: 0,
-    });
-  };
-
-  return (
-    <div className="mt-2 rounded border border-border bg-secondary/30 px-3 py-2 text-xs">
-      {scanning && (
-        <p className="text-muted-foreground flex items-center gap-2">
-          <LoadingSpinner size="sm" inline /> Testing window trims across episodes, this can take a minute...
-        </p>
-      )}
-      {!scanning && scanError && (
-        <div className="flex flex-wrap items-center gap-3">
-          <p className="text-destructive">{scanError}</p>
-          <button type="button" className={`px-2 py-1 rounded ${ghostBtn}`} onClick={rescan}>
-            Rescan
-          </button>
-          <button type="button" className={`px-2 py-1 rounded ${ghostBtn}`} onClick={onClose}>
-            Close
-          </button>
-        </div>
-      )}
-      {!scanning && !scanError && ready && (
-        <div className="space-y-2">
-          {alreadyOptimal ? (
-            <p>
-              <span className="px-2 py-0.5 rounded font-medium bg-green-500/20 text-green-600 dark:text-green-400">
-                Already optimal
-              </span>
-              <span className="ml-2 text-muted-foreground">
-                No trim scored higher than the current window.
-              </span>
-            </p>
-          ) : (
-            <div className="grid max-w-sm grid-cols-[auto_1fr_1fr] gap-x-4 gap-y-0.5 font-mono">
-              <span />
-              <span className="font-sans text-muted-foreground">Current</span>
-              <span className="font-sans text-muted-foreground">Proposed</span>
-              <span className="font-sans text-muted-foreground">Start</span>
-              <span>{baselineWindow ? `${baselineWindow.startS.toFixed(2)}s` : '--'}</span>
-              <span>{proposedStartS.toFixed(2)}s</span>
-              <span className="font-sans text-muted-foreground">End</span>
-              <span>{baselineWindow ? `${baselineWindow.endS.toFixed(2)}s` : '--'}</span>
-              <span>{proposedEndS.toFixed(2)}s</span>
-              <span className="font-sans text-muted-foreground">Score</span>
-              <span title={baselineMeanPeakScore == null
-                ? 'The current window is outside the capture bounds, so it was not scored'
-                : undefined}
-              >
-                {baselineMeanPeakScore != null ? baselineMeanPeakScore.toFixed(3) : '--'}
-              </span>
-              <span>
-                {meanPeakScore.toFixed(3)}
-                {scoreDelta != null && (
-                  <span className={`ml-1 ${scoreDelta >= 0
-                    ? 'text-green-600 dark:text-green-400'
-                    : 'text-amber-600 dark:text-amber-400'}`}
-                  >
-                    {scoreDelta >= 0 ? '+' : ''}{scoreDelta.toFixed(3)}
-                  </span>
-                )}
-              </span>
-            </div>
-          )}
-          {(perEpisode?.length ?? 0) > 0 && (
-            <p className="text-muted-foreground">
-              Per episode:{' '}
-              {perEpisode!.map((e) => (
-                <span key={e.episodeId} className="mr-2 font-mono">
-                  {e.episodeId.slice(0, 8)} {e.peakScore.toFixed(3)}
-                </span>
-              ))}
-            </p>
-          )}
-          <div className="flex flex-wrap items-center gap-2">
-            <button type="button" className={`px-2 py-1 rounded ${ghostBtn}`} onClick={rescan}>
-              Rescan
-            </button>
-            <button type="button" className={`px-2 py-1 rounded ${ghostBtn}`} onClick={onClose}>
-              Discard
-            </button>
-            {!alreadyOptimal && (
-              <button
-                type="button"
-                className={`px-2 py-1 rounded ${primaryBtn} disabled:opacity-50`}
-                onClick={() => {
-                  setApplyError(null);
-                  applyMutation.mutate({ startS: proposedStartS, endS: proposedEndS });
-                }}
-                disabled={applyMutation.isPending}
-              >
-                {applyMutation.isPending ? 'Applying...' : 'Apply'}
-              </button>
-            )}
-          </div>
-          {applyError && <p className="text-destructive">{applyError}</p>}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Maximum episodes a user may select for the cross-episode scan (server cap).
-const CROSS_EPISODE_MAX = 5;
-const CROSS_EPISODE_MIN = 2;
-
-interface CueCrossEpisodeScanModalProps {
-  slug: string;
-  captureMinSeconds: number;
-  captureMaxSeconds: number;
-  captureMaxIntroSeconds: number;
-  captureMaxOutroSeconds: number;
-  onClose: () => void;
-  onSaved: () => void;
-}
-
-function CueCrossEpisodeScanModal({
-  slug,
-  captureMinSeconds,
-  captureMaxSeconds,
-  captureMaxIntroSeconds,
-  captureMaxOutroSeconds,
-  onClose,
-  onSaved,
-}: CueCrossEpisodeScanModalProps) {
-  useEscape(onClose);
-  const queryClient = useQueryClient();
-  const [pickerPage, setPickerPage] = useState(0);
-  // Selected episodes in click order (first = target). Full objects, not ids,
-  // so title/duration survive paging away from the page they were picked on.
-  const [selected, setSelected] = useState<Episode[]>([]);
-  // Phase: picker -> results (scanning/ready/error handled in scanQuery state).
-  const [phase, setPhase] = useState<'picker' | 'results'>('picker');
-  // Seed for CueMarkModal when a candidate's "Make template" is clicked.
-  const [seed, setSeed] = useState<CrossEpisodeCandidate | null>(null);
-
-  const pickerQuery = useQuery({
-    queryKey: ['cue-template-picker', slug, pickerPage],
-    queryFn: () =>
-      getEpisodes(slug, {
-        limit: PICKER_PAGE_SIZE,
-        offset: pickerPage * PICKER_PAGE_SIZE,
-        status: 'completed',
-        sortBy: 'published',
-        sortDir: 'desc',
-      }),
-    enabled: !!slug,
-  });
-
-  const pickerEpisodes = (pickerQuery.data?.episodes ?? [])
-    .filter((ep) => ep.hasOriginalAudio !== false);
-  const pickerTotal = pickerQuery.data?.total ?? 0;
-  const pickerTotalPages = Math.max(1, Math.ceil(pickerTotal / PICKER_PAGE_SIZE));
-
-  const selectedIds = selected.map((ep) => ep.id);
-  const targetEp = selected[0];
-
-  // React Query scan: enabled once user advances to results phase.
-  // Polling stops when status is no longer 'scanning'.
-  const scanQueryKey = ['cue-cross-episode-scan', slug, selectedIds];
-  const scanQuery = useQuery<CrossEpisodeScanResponse>({
-    queryKey: scanQueryKey,
-    queryFn: () => crossEpisodeScan(slug, selectedIds),
-    enabled: phase === 'results' && selected.length >= CROSS_EPISODE_MIN,
-    staleTime: Infinity,
-    refetchInterval: (q) =>
-      q.state.data?.status === 'scanning' ? 3000 : false,
-  });
-
-  const scanData = scanQuery.data;
-  const scanning = phase === 'results' && (scanQuery.isLoading || scanData?.status === 'scanning');
-  const scanError = scanData?.status === 'error'
-    ? (scanData.error || 'Scan failed.')
-    : (scanQuery.error ? 'Scan failed. Try again.' : null);
-  const candidates: CrossEpisodeCandidate[] = scanData?.candidates ?? [];
-
-  const toggleEpisode = (ep: Episode) => {
-    setSelected((prev) => {
-      if (prev.some((p) => p.id === ep.id)) return prev.filter((p) => p.id !== ep.id);
-      if (prev.length >= CROSS_EPISODE_MAX) return prev;
-      return [...prev, ep];
-    });
-  };
-
-  // Force a fresh server-side run; same fetchQuery idiom as
-  // CueCandidatesSection's rescan, so repeated clicks always refetch.
-  const rescan = () =>
-    queryClient.fetchQuery({
-      queryKey: scanQueryKey,
-      queryFn: () => crossEpisodeScan(slug, selectedIds, true),
-      staleTime: 0,
-    });
-
-  // Picker phase
-  if (phase === 'picker') {
-    return (
-      <div className={modalBackdrop} onClick={onClose}>
-        <div
-          role="dialog"
-          aria-modal="true"
-          aria-label="Find cues across episodes"
-          className={`${modalPanel} w-full max-w-2xl p-5 max-h-[85vh] flex flex-col`}
-          onClick={(e) => e.stopPropagation()}
-        >
-          <div className="flex items-start justify-between mb-3">
-            <div>
-              <h3 className="text-base font-semibold">Find cues across episodes</h3>
-              <p className="text-xs text-muted-foreground">
-                Select {CROSS_EPISODE_MIN}-{CROSS_EPISODE_MAX} episodes. Results are shown in the first selected episode's time.
-              </p>
-            </div>
-            <button type="button" className="text-muted-foreground hover:text-foreground" onClick={onClose} aria-label="Close">
-              <X size={18} />
-            </button>
-          </div>
-
-          <div className="flex-1 overflow-y-auto border border-border rounded">
-            {pickerQuery.isLoading && <div className="p-4"><LoadingSpinner size="sm" /></div>}
-            {pickerQuery.error && <p className="p-3 text-sm text-destructive">Could not load episodes.</p>}
-            {!pickerQuery.isLoading && pickerEpisodes.length === 0 && (
-              <p className="p-3 text-sm text-muted-foreground">No episodes with original audio found.</p>
-            )}
-            {pickerEpisodes.length > 0 && (
-              <ul className="divide-y divide-border">
-                {pickerEpisodes.map((ep) => {
-                  const checked = selectedIds.includes(ep.id);
-                  const atMax = !checked && selectedIds.length >= CROSS_EPISODE_MAX;
-                  const rank = selectedIds.indexOf(ep.id);
-                  return (
-                    <li key={ep.id}>
-                      <label
-                        className={`flex items-start gap-3 px-3 py-2 cursor-pointer select-none ${atMax ? 'opacity-50 cursor-not-allowed' : 'hover:bg-muted/50'}`}
-                      >
-                        <input
-                          type="checkbox"
-                          checked={checked}
-                          disabled={atMax}
-                          onChange={() => toggleEpisode(ep)}
-                          className="mt-0.5 shrink-0"
-                          aria-label={`Select episode ${ep.title}`}
-                        />
-                        <div className="flex-1 min-w-0">
-                          <p className="text-sm font-medium truncate">
-                            {ep.title}
-                            {rank === 0 && (
-                              <span className="ml-2 px-1.5 py-0.5 text-xs rounded font-medium bg-primary/20 text-primary align-middle">
-                                target
-                              </span>
-                            )}
-                          </p>
-                          <p className="text-xs text-muted-foreground">
-                            {ep.published ? new Date(ep.published).toLocaleDateString() : 'unknown date'}
-                            {typeof ep.duration === 'number' && ep.duration > 0
-                              ? ` - ${Math.round(ep.duration / 60)} min` : ''}
-                          </p>
-                        </div>
-                      </label>
-                    </li>
-                  );
-                })}
-              </ul>
-            )}
-          </div>
-
-          {pickerTotalPages > 1 && (
-            <div className="flex items-center justify-between mt-3 text-sm">
-              <button
-                type="button"
-                className={`px-2 py-1 rounded ${ghostBtn} disabled:opacity-50`}
-                onClick={() => setPickerPage((p) => Math.max(0, p - 1))}
-                disabled={pickerPage === 0}
-              >
-                Prev
-              </button>
-              <span className="text-muted-foreground">
-                Page {pickerPage + 1} / {pickerTotalPages}
-              </span>
-              <button
-                type="button"
-                className={`px-2 py-1 rounded ${ghostBtn} disabled:opacity-50`}
-                onClick={() => setPickerPage((p) => Math.min(pickerTotalPages - 1, p + 1))}
-                disabled={pickerPage + 1 >= pickerTotalPages}
-              >
-                Next
-              </button>
-            </div>
-          )}
-
-          <div className="mt-3 flex items-center justify-between gap-3">
-            <div className="text-xs text-muted-foreground">
-              {selected.length === 0 && 'Select at least 2 episodes.'}
-              {selected.length === 1 && 'Select 1 more episode.'}
-              {selected.length >= CROSS_EPISODE_MIN && (
-                <>
-                  {selected.length} selected{selected.length === CROSS_EPISODE_MAX ? ' (max)' : ''}
-                  {targetEp && (
-                    <> - results on: <span className="font-medium text-foreground truncate max-w-[180px] inline-block align-bottom">{targetEp.title}</span></>
-                  )}
-                </>
-              )}
-            </div>
-            <button
-              type="button"
-              className={`px-3 py-1.5 rounded ${primaryBtn} text-sm disabled:opacity-50`}
-              disabled={selected.length < CROSS_EPISODE_MIN}
-              onClick={() => setPhase('results')}
-            >
-              Scan
-            </button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  // Results phase
-  return (
-    <>
-      <div className={modalBackdrop} onClick={onClose}>
-        <div
-          role="dialog"
-          aria-modal="true"
-          aria-label="Cross-episode scan results"
-          className={`${modalPanel} w-full max-w-2xl p-5 max-h-[85vh] flex flex-col`}
-          onClick={(e) => e.stopPropagation()}
-        >
-          <div className="flex items-start justify-between mb-3">
-            <div>
-              <h3 className="text-base font-semibold">Cross-episode scan</h3>
-              {targetEp && (
-                <p className="text-xs text-muted-foreground truncate max-w-xl">
-                  Results on: {targetEp.title}
-                </p>
-              )}
-            </div>
-            <button type="button" className="text-muted-foreground hover:text-foreground" onClick={onClose} aria-label="Close">
-              <X size={18} />
-            </button>
-          </div>
-
-          <div className="flex flex-wrap gap-2 mb-3">
-            <button
-              type="button"
-              className={`px-3 py-1.5 rounded ${ghostBtn} text-sm`}
-              onClick={() => setPhase('picker')}
-            >
-              Change episodes
-            </button>
-            {!scanning && (
-              <button
-                type="button"
-                className={`px-3 py-1.5 rounded ${ghostBtn} text-sm`}
-                onClick={() => rescan()}
-              >
-                Rescan
-              </button>
-            )}
-          </div>
-
-          {scanning && (
-            <p className="text-sm text-muted-foreground flex items-center gap-2 mb-3">
-              <LoadingSpinner size="sm" inline /> Scanning audio, this can take a minute...
-            </p>
-          )}
-          {!scanning && scanError && (
-            <p className="text-sm text-destructive mb-3">{scanError}</p>
-          )}
-          {!scanning && !scanError && scanData?.status === 'ready' && candidates.length === 0 && (
-            <p className="text-sm text-muted-foreground">No recurring segments found.</p>
-          )}
-
-          {candidates.length > 0 && (
-            <ul className="flex-1 overflow-y-auto divide-y divide-border border border-border rounded">
-              {candidates.map((c) => (
-                <li key={`${c.start}-${c.end}`} className="flex items-center gap-3 px-3 py-2 text-sm">
-                  <div className="flex-1 min-w-0">
-                    <span className="font-mono text-sm">
-                      {formatTimestamp(c.start)} - {formatTimestamp(c.end)}
-                    </span>
-                    <span className="ml-2 text-xs text-muted-foreground">
-                      {(c.end - c.start).toFixed(2)}s
-                    </span>
-                    {c.episodeMatches != null && (
-                      <span className="ml-2 px-1.5 py-0.5 text-xs rounded font-medium bg-blue-500/20 text-blue-600 dark:text-blue-400">
-                        matched in {c.episodeMatches} ep{c.episodeMatches === 1 ? '' : 's'}
-                      </span>
-                    )}
-                  </div>
-                  <button
-                    type="button"
-                    className={`shrink-0 px-3 py-1.5 rounded ${primaryBtn} text-xs`}
-                    onClick={() => setSeed(c)}
-                  >
-                    Make template
-                  </button>
-                </li>
-              ))}
-            </ul>
-          )}
-        </div>
-      </div>
-
-      {seed && targetEp && (
-        <CueMarkModal
-          podcastSlug={slug}
-          episodeId={targetEp.id}
-          episodeTitle={targetEp.title}
-          episodeDuration={targetEp.duration ?? 0}
-          initialStart={seed.start}
-          initialEnd={seed.end}
-          captureMinSeconds={captureMinSeconds}
-          captureMaxSeconds={captureMaxSeconds}
-          captureMaxIntroSeconds={captureMaxIntroSeconds}
-          captureMaxOutroSeconds={captureMaxOutroSeconds}
-          onClose={() => setSeed(null)}
-          onSaved={onSaved}
-          onFinalSave={() => setSeed(null)}
-        />
-      )}
-    </>
   );
 }
